Move `RenderingTestbed` into `testlib` module
