+++ api-tests/src/test/kotlin/io/spine/protodata/context/MemberSpec.kt
@@ -31,7 +31,7 @@ import io.kotest.matchers.string.shouldEndWith
 import io.spine.protodata.ast.EnumInFile
 import io.spine.protodata.ast.MessageInFile
 import io.spine.protodata.ast.ServiceInFile
-import io.spine.protodata.backend.RenderingTestbed
+import io.spine.protodata.testing.RenderingTestbed
 import io.spine.protodata.render.Renderer
 import io.spine.protodata.render.SourceFileSet
 import io.spine.protodata.testing.PipelineSetup
